fix gcc warning by making integer casts explicit
git-svn-id: http://vtp.googlecode.com/svn/trunk@2182 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/core/Globe.cpp
@@ -718,8 +718,8 @@ void IcoGlobe::EstimateTesselation(int iTriangleCount)
 		// triangles.  Find what frequency most closely matches the desired
 		// triangle count.
 		double exact = sqrt(per_face);
-		int iLess = floor(exact);
-		int iMore = ceil(exact);
+		int iLess = (int) floor(exact);
+		int iMore = (int) ceil(exact);
 		if ((iMore*iMore - per_face) < (per_face - iLess*iLess))
 			m_freq = iMore;
 		else
